gtm-prod1/bot/requirements_supabase.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.9
redis==5.0.1 